requirements.txt
Flask==1.1.1
celery==3.1.25
Flask_RESTful==0.3.8
Werkzeug==0.16.0
requests==2.22.0
psycopg2==2.9.1
PyJWT==2.1.0
